Separate equity linearity metrics from node wrappers
