+++ qmtl/transforms/__init__.py
@@ -48,10 +48,12 @@ from .execution_velocity_hazard import (
     execution_velocity_hazard,
 )
 from .order_book_inertia import obii_from_survival, order_book_inertia
-from .equity_linearity import (
+from .linearity_metrics import (
     equity_linearity_metrics,
-    equity_linearity_from_history_node,
     equity_linearity_metrics_v2,
+)
+from .equity_linearity import (
+    equity_linearity_from_history_node,
     equity_linearity_v2_from_history_node,
 )
